Use git submodule to include pthread-win32; update to latest master
Using git submodules makes it easier to figure out the
exact version of pthread-win32 being used, as well as
simplifying updating to a new version down the line.

The local changes made to the pthread.h header have
not been reapplied, and instead the necessary
preprocessor macros are now defined in the premake
build specification.